activation: refactor CrossVersionProps.scala
